requirements.txt
python-dotenv
pyrogram
psutil
